Users that have taken the screener can now view previous results or retake the screener
